Do not display in ingame display but in map display

+++ src/inGame/display.cpp
@@ -10,10 +10,12 @@ void ef::InGame::display(std::shared_ptr<Bpixelarray> screen,
       bunny_set_geometry(&screen->GetClip()->buffer, BGY_LINES, (t_bunny_vertex_array *)&screen->lineVec, NULL);
       screen->lineVec.length = 0;
 
+      /*
       for (int i = 0; i < (int)myUnits.size(); i++)
-	myUnits[i]->display(screen, ef::AcuPos{cameraPos.x, cameraPos.y, 0});
+	myUnits[i]->display(screen, ef::AcuPos{(double)cameraPos.x, (double)cameraPos.y, 0});
       for (int i = 0; i < (int)enemyUnits.size(); i++)
-	enemyUnits[i]->display(screen, ef::AcuPos{cameraPos.x, cameraPos.y, 0});
+	enemyUnits[i]->display(screen, ef::AcuPos{(double)cameraPos.x, (double)cameraPos.y, 0});
+      */
     }
   else
     {
